Re-generate poetry.lock
Ensure the lock file is up to date, without upgrading dependencies.
Poetry does not currently offer an interface for this, but it is
possible by adding and removing an unrelated dependency. We use
`insecure-package` from the pyup project:

  $ poetry add insecure-package
  $ poetry remove insecure-package

The diff shows that platform markers were added for cryptography and
cffi. Note that cryptography is the package being upgraded by this
dependabot PR.

(The remaining items in the diff are essentially noise: changes in the
list of downloaded wheels, and an equivalent reordering of conditions
inside the colorama marker. The latter may be due to a harmless Poetry
bug, and happens to me every time it touches the lock file.)

Without the markers for cryptography and cffi, the build fails on macOS:

  https://github.com/cjolowicz/poetry-merge-lock/pull/41/checks?check_run_id=629974751

The build fails after the Nox session exports the lock file to
requirements.txt format, and installs it using pip. This is the
error message from pip:

  ERROR: In --require-hashes mode, all requirements must have their versions pinned with ==. These do not:
      pycparser from
      https://files.pythonhosted.org/packages/ae/e7/d9c3a176ca4b02024debf82342dab36efadfc5776f9c8db077e8f6e71821/pycparser-2.20-py2.py3-none-any.whl#sha256=7582ad22678f0fcd81102833f60ef8d0e57288b6b5fb00323d101be910e35705
      (from cffi==1.14.0->-r
      /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/tmpxx2rvg5y/requirements.txt
      (line 13))